SECS v0.2 - Tons of internal changes, redid Queries, and more...
This is almost a whole rewrite. Seriously. There's a lot of improvements here.

First off, queries work almost on-par with Bevy's queries now - you can query from multiple components, instead of it being basically a relabelled archetype. Queries can also now give you the entity's ID for a component directly. There's no change detection/state, nor filters, but those will hopefully come in the future

Entity spawning has been redone. Instead of being spawned with the `EntityBuilder`, they're now spawned with `World.spawn()` and a tuple of components.

The internal storage design has changed. Moving items in Rust may-or-may-not actually move them in physical memory; this meant every time system parameters took ownership of archetypes and other data in the world, those items may have been moved in memory, which would be very costly. Everything now uses `Rc`s and `RefCell`s, so borrowing is determined at runtime, and the `Rc` should keep items in-place.

The world's storage has been separated from the actual world. In the future, I may implement a storage trait that allows for custom storage backends. I tried to add it this time, and it required adding generics to *everything* that interacted with the world... which then meant adding generics to everything that used those types, and so on... the code got excessively messy. This is a step towards separating the storage and providing an API for that, but it's not yet there.